+++ src/modules/clan-capital/traps/log-trap/index.ts
@@ -1,0 +1,4 @@
+import rawData from '@/data/clan-capital/traps/log-trap.json';
+import { ClanCapitalTrap } from '@/types';
+
+export const logTrapData: ClanCapitalTrap = rawData as unknown as ClanCapitalTrap;