+++ src/modules/clan-capital/traps/mega-mine/index.ts
@@ -1,0 +1,4 @@
+import rawData from '@/data/clan-capital/traps/mega-mine.json';
+import { ClanCapitalTrap } from '@/types';
+
+export const megaMineData: ClanCapitalTrap = rawData as unknown as ClanCapitalTrap;