+++ src/modules/clan-capital/traps/mine/index.ts
@@ -1,0 +1,4 @@
+import rawData from '@/data/clan-capital/traps/mine.json';
+import { ClanCapitalTrap } from '@/types';
+
+export const mineData: ClanCapitalTrap = rawData as unknown as ClanCapitalTrap;