+++ src/modules/clan-capital/traps/spear-trap/index.ts
@@ -1,0 +1,4 @@
+import rawData from '@/data/clan-capital/traps/spear-trap.json';
+import { ClanCapitalTrap } from '@/types';
+
+export const spearTrapData: ClanCapitalTrap = rawData as unknown as ClanCapitalTrap;