+++ src/modules/clan-capital/traps/zap-trap/index.ts
@@ -1,0 +1,4 @@
+import rawData from '@/data/clan-capital/traps/zap-trap.json';
+import { ClanCapitalTrap } from '@/types';
+
+export const zapTrapData: ClanCapitalTrap = rawData as unknown as ClanCapitalTrap;